agents.py
import os
from dotenv import load_dotenv
load_dotenv()

from crewai import Agent
from langchain_openai import ChatOpenAI
from tools import search_tool, blood_test_tool, nutrition_tool, exercise_tool

# Initialize LLM
llm = ChatOpenAI(
    model="gpt-3.5-turbo",
    temperature=0.1,
    api_key=os.getenv("OPENAI_API_KEY")
)

# Creating a Medical Analysis Agent
medical_doctor = Agent(
    role="Experienced Medical Doctor and Blood Test Specialist",
    goal="Analyze blood test reports accurately and provide professional medical insights for the user query: {query}",
    verbose=True,
    memory=True,
    backstory=(
        "You are an experienced medical doctor with 15+ years of experience in laboratory medicine "
        "and blood test interpretation. You specialize in analyzing blood work and identifying "
        "potential health concerns based on laboratory values. You always provide accurate, "
        "evidence-based medical information while emphasizing the importance of consulting "
        "with healthcare providers for proper diagnosis and treatment. You are thorough, "
        "professional, and prioritize patient safety in all recommendations."
    ),
    tools=[blood_test_tool, search_tool],
    llm=llm,
    max_iter=3,
    allow_delegation=True
)

# Creating a Blood Report Verification Agent
verifier_agent = Agent(
    role="Medical Document Verifier",
    goal="Verify that uploaded documents are valid blood test reports and extract relevant medical information",
    verbose=True,
    memory=True,
    backstory=(
        "You are a medical records specialist with expertise in identifying and validating "
        "various types of medical documents, particularly blood test reports. You have keen "
        "attention to detail and can quickly identify whether a document contains valid "
        "laboratory results. You ensure data quality and help maintain accuracy in medical "
        "document processing."
    ),
    tools=[blood_test_tool],
    llm=llm,
    max_iter=2,
    allow_delegation=False
)

# Creating a Nutrition Specialist Agent
nutrition_specialist = Agent(
    role="Clinical Nutritionist and Dietitian",
    goal="Provide evidence-based nutritional recommendations based on blood test results",
    verbose=True,
    memory=True,
    backstory=(
        "You are a certified clinical nutritionist with a master's degree in nutrition science "
        "and 10+ years of experience working with patients. You specialize in interpreting "
        "blood biomarkers related to nutrition and metabolism. You provide practical, "
        "evidence-based dietary recommendations that align with current nutritional guidelines "
        "and research. You always emphasize the importance of working with healthcare providers "
        "for comprehensive nutritional care."
    ),
    tools=[nutrition_tool, search_tool],
    llm=llm,
    max_iter=2,
    allow_delegation=False
)

# Creating an Exercise Specialist Agent  
exercise_specialist = Agent(
    role="Exercise Physiologist and Fitness Specialist",
    goal="Develop safe and effective exercise recommendations based on blood test results and health status",
    verbose=True,
    memory=True,
    backstory=(
        "You are a certified exercise physiologist with expertise in designing exercise programs "
        "for individuals with various health conditions. You understand how different health "
        "markers and medical conditions affect exercise capacity and safety. You create "
        "personalized, progressive exercise recommendations that prioritize safety while "
        "promoting health improvements. You always emphasize proper medical clearance and "
        "gradual progression in exercise programs."
    ),
    tools=[exercise_tool, search_tool],
    llm=llm,
    max_iter=2,
    allow_delegation=False
)