requirements.txt
fastapi
uvicorn[standard]
crewai
crewai-tools
python-dotenv
python-multipart
PyPDF2
langchain-community
openai
celery
redis
sqlalchemy
alembic
pydantic
onnxruntime
chromadb
tokenizers 
tqdm
